+++ src/evidently/metrics/data_integrity/column_missing_values_metric.py
@@ -1,0 +1,192 @@
+from typing import Any
+from typing import Dict
+from typing import List
+from typing import Optional
+
+import dataclasses
+import numpy as np
+import pandas as pd
+
+from evidently.metrics.base_metric import InputData
+from evidently.metrics.base_metric import Metric
+from evidently.model.widget import BaseWidgetInfo
+from evidently.renderers.base_renderer import MetricRenderer
+from evidently.renderers.base_renderer import default_renderer
+from evidently.renderers.html_widgets import header_text
+from evidently.renderers.html_widgets import table_data
+
+
+@dataclasses.dataclass
+class ColumnMissingValues:
+    """Statistics about missing values in a column"""
+
+    # count of rows in the column
+    number_of_rows: int
+    # set of different missed values in the column
+    different_missing_values: Dict[Any, int]
+    # number of different missed values in the column
+    number_of_different_missing_values: int
+    # count of missed values in the column
+    number_of_missing_values: int
+    # share of missed values in the column
+    share_of_missing_values: float
+
+
+@dataclasses.dataclass
+class ColumnMissingValuesMetricResult:
+    column_name: str
+    current: ColumnMissingValues
+    reference: Optional[ColumnMissingValues] = None
+
+
+class ColumnMissingValuesMetric(Metric[ColumnMissingValuesMetricResult]):
+    """Count missing values in a column.
+
+    Missing value is a null or NaN value.
+
+    Calculate an amount of null-like values kinds and count for such values.
+    NA-types like numpy.NaN, pandas.NaT are counted as one type.
+
+    You can set you own null-line values list with `values` parameter.
+    Value None in the list means that Pandas null values will be included in the calculation.
+
+    If `replace` parameter is False - add defaults to user's list.
+    If `replace` parameter is True - use values from `null_values` list only.
+    """
+
+    # default missing values list
+    DEFAULT_VALUES = ["", np.inf, -np.inf, None]
+    values: frozenset
+    column_name: str
+
+    def __init__(self, column_name: str, values: Optional[list] = None, replace: bool = True) -> None:
+        self.column_name = column_name
+
+        if values is None:
+            # use default null-values list if we have no user-defined null values
+            values = self.DEFAULT_VALUES
+
+        elif not replace:
+            # add default nulls to user-defined nulls list
+            values = self.DEFAULT_VALUES + values
+
+        # use frozenset because metrics parameters should be immutable/hashable for deduplication
+        self.values = frozenset(values)
+
+    def _calculate_null_values_stats(self, column: pd.Series) -> ColumnMissingValues:
+        different_missing_values = {value: 0 for value in self.values}
+        number_of_missing_values = 0
+
+        number_of_rows = len(column)
+
+        # iterate by each value in custom null-values list and check the value in a column
+        for value in self.values:
+            if value is None:
+                # check all pandas null-types like numpy.NAN, pandas.NA, pandas.NaT, etc
+                missing_values = column.isnull().sum()
+
+            else:
+                missing_values = (column == value).sum()
+
+            if missing_values > 0:
+                # increase overall counter
+                number_of_missing_values += missing_values
+                # increase by-null-value counter
+                different_missing_values[value] += missing_values
+
+        share_of_missing_values = number_of_missing_values / number_of_rows
+
+        number_of_different_missing_values = sum(
+            [1 for null_value in different_missing_values if different_missing_values[null_value] > 0]
+        )
+
+        return ColumnMissingValues(
+            different_missing_values=different_missing_values,
+            number_of_different_missing_values=number_of_different_missing_values,
+            number_of_missing_values=number_of_missing_values,
+            share_of_missing_values=share_of_missing_values,
+            number_of_rows=number_of_rows,
+        )
+
+    def calculate(self, data: InputData) -> ColumnMissingValuesMetricResult:
+        if not self.values:
+            raise ValueError("Missed values list should not be empty.")
+
+        if self.column_name not in data.current_data:
+            raise ValueError(f"Column {self.column_name} is not in current data.")
+
+        current_null_values = self._calculate_null_values_stats(data.current_data[self.column_name])
+
+        if data.reference_data is None:
+            reference_null_values: Optional[ColumnMissingValues] = None
+
+        else:
+            if self.column_name not in data.reference_data:
+                raise ValueError(f"Column {self.column_name} is not in reference data.")
+
+            reference_null_values = self._calculate_null_values_stats(data.reference_data[self.column_name])
+
+        return ColumnMissingValuesMetricResult(
+            column_name=self.column_name,
+            current=current_null_values,
+            reference=reference_null_values,
+        )
+
+
+@default_renderer(wrap_type=ColumnMissingValuesMetric)
+class ColumnMissingValuesMetricRenderer(MetricRenderer):
+    def render_json(self, obj: ColumnMissingValuesMetric) -> dict:
+        return dataclasses.asdict(obj.get_result())
+
+    @staticmethod
+    def _get_table_stat(dataset_name: str, stats: ColumnMissingValues) -> BaseWidgetInfo:
+        data = []
+
+        for missed_value, count_of_missed in stats.different_missing_values.items():
+            percent_of_missed = round(count_of_missed / stats.number_of_rows * 100, 3)
+            value = f"{count_of_missed} ({percent_of_missed}%)"
+
+            if missed_value is None:
+                missed_value_str = "Pandas and Numpy null, NA, NaN, etc."
+
+            elif not missed_value:
+                missed_value_str = "Empty string"
+
+            else:
+                missed_value_str = str(missed_value)
+
+            data.append((missed_value_str, value))
+
+        matched_stat_headers = ["Missed Value", "Count"]
+        return table_data(
+            title=f"{dataset_name.capitalize()}: Missed values",
+            column_names=matched_stat_headers,
+            data=data,
+        )
+
+    @staticmethod
+    def _get_details_missing_values_info(dataset_name: str, stats: ColumnMissingValues) -> BaseWidgetInfo:
+        percents = round(stats.share_of_missing_values * 100, 3)
+        return header_text(
+            label=f"In {dataset_name} dataset {stats.number_of_missing_values}({percents}%) missed values."
+        )
+
+    def render_html(self, obj: ColumnMissingValuesMetric) -> List[BaseWidgetInfo]:
+        metric_result = obj.get_result()
+
+        result = [
+            header_text(label=f"Missing values in column '{metric_result.column_name}'"),
+            self._get_details_missing_values_info(dataset_name="current", stats=metric_result.current),
+        ]
+
+        if metric_result.reference is not None:
+            result.append(
+                self._get_details_missing_values_info(dataset_name="reference", stats=metric_result.reference),
+            )
+
+        result.append(self._get_table_stat(dataset_name="current", stats=metric_result.reference))
+
+        if metric_result.reference is not None:
+            result.append(self._get_table_stat(dataset_name="reference", stats=metric_result.reference))
+
+        return result
